moved code, exported some helper functions
